requirements.txt
requests
minio
pandas
sqlalchemy
psycopg2-binary
numpy
